Conflicts:
	perl/openprint/Object.pm

+++ perl/openprint/employee_inventory.pm
@@ -1368,13 +1368,16 @@ sub manifest {
 				my $total_qty = 0;
 				# Save data for the rest of the contents
 				foreach my $C ( $Manifest->Contents( 'type_id' => $Type->id() ) ) {
-					if ( exists $param{"qty_lbs-$$Type{id}-$$C{id}"} ) {
+					my $checked_out = openprint::PaperInventory->find('skid_id'=>$C->Skid()->id(), 'paper_id'=>undef, 'comment_like'=>'Checked out%' ) ? 1 : 0; 
+					if ( exists $param{"qty_lbs-$$Type{id}-$$C{id}"} and ( $C->quantity() != $param{"qty_lbs-$$Type{id}-$$C{id}"} ) ) {
 						$variable{'error'} .= $C->save({ 'quantity'	=> sprintf('%d', $param{"qty_lbs-$$Type{id}-$$C{id}"}) });
+						if ( ! $checked_out ) {
+							save_inventory( $C->Skid(), $Paper, $C->quantity(), sprintf('Inventory adjusted from manifest %1$s.', $Manifest->name() ) );
+						} # end if
 					} # end if
 					$total_qty += $C->quantity();
-					save_inventory( $C->Skid(), $Paper, $C->quantity(), sprintf('Inventory adjusted from manifest %1$s.', $Manifest->name() ) );
 					#if ( $Project and ( $param{"allocate-$$Type{id}"} eq 'Specific' ) ) {
-					if ( $Project ) {
+					if ( $Project and ! $checked_out ) {
 						my $PA = openprint::PaperAllocation->find_one('skid_id'=>$C->skid_id());
 						if ( ! $PA ) {
 							$Paper->allocate( $C->Skid(), $Project->id(), $C->quantity(), $Paper->type() eq 'Roll' ? 'lbs' : 'sheets' );
@@ -1385,10 +1388,6 @@ sub manifest {
 						} elsif ( $PA->project_id() != $Project->id() ) {
 							$variable{'information'} .= sprintf('Skid <a href="/employee/inventory/skid_details.html?skid_id=%1$d">%1$d</a> already allocated to docket <a href="/employee/project/view.html?ProjectIndex=%2$d">%3$d</a>.<br/>', $C->skid_id(), $PA->project_id(), $PA->docket() );
 						} # end if
-					} # end if
-					if ( openprint::PaperInventory->find('skid_id'=>$C->Skid()->id(), 'paper_id'=>undef, 'comment_like'=>'Checked out%' ) ) {
-						# If the stock has already been checked out, add a subtraction to keep counts in line.
-						save_inventory( $C->Skid(), $Paper, -1*$C->quantity(), 'Automatic checkout after manifest inventory update.' );
 					} # end if
 				} # end foreach tag_id
 
@@ -1602,7 +1601,11 @@ sub purchase_order_view {
 
 	my $Me = new openprint::User( $session{'user_id'} );
 	my $PO = new openprint::PurchaseOrder( $param{'po_id'} );
-
+	if ( ! $PO->id() ) {
+		$variable{'error'} .= 'Invalid PO # given: ' . $param{'po_id'}.'<br/>';
+		$variable{'PurchaseOrder'} = $PO;
+		return;
+	} # end if
 	if ( $param{'btnFunction'} eq 'Delete' ) {
 		$variable{'error'} .= $PO->delete();
 		if ( ! $variable{'error'} ) {
@@ -1669,10 +1672,22 @@ sub purchase_order_view {
 			} # end foreach
 			$New->save();
 			$variable{'information'} .= 'PO ' . $PO->id() . ' copied to PO ' . $New->id() .'<br/>';
+			my $L = new openprint::PurchaseOrder_Log();
+			$L->save({
+					'user_id'	=>	$session{'user_id'},
+					'po_id'		=>	$New->id(),
+					'reason'	=>	'Copied from PO '. $PO->id(),
+					});
+			$L = new openprint::PurchaseOrder_Log();
+			$L->save({
+					'user_id'	=>	$session{'user_id'},
+					'po_id'		=>	$PO->id(),
+					'reason'	=>	'Copied to PO '. $New->id(),
+					});
 			$PO = $New;
 		} # end if
 		if ( ! $PO->authorized() ) {
-			if ( $PO->total() < $Me->purchasing_limit() ) {
+			if ( sets::isin( $session{'user_type'}, ['A'] ) or ( $PO->total() < $Me->purchasing_limit() ) ) {
 				$variable{'error'} .= $PO->save({
 						'authorized'	=> 1,
 						'authorized_on'	=> 'NOW()',
@@ -1705,7 +1720,7 @@ sub purchase_order_view {
 						'type_id'		=>	$param{'type_id-'.$content_id},
 						});
 				$types{$C->Type()->name()} = 1;
-				if ( $C->docket() ) {
+				if ( $C->docket() and ! ( $C->docket() =~ /\D/ ) ) {
 					foreach my $P ( openprint::Project->find('docket'=>$C->docket()) ) {
 						$P->add_to_log( @session{'company_id','user_id'}, 
 								sprintf('<a href="/employee/inventory/purchase_order_view.html?po_id=%1$d">%2$s%3$s %4$s ordered on PO%1$d</a>',
@@ -1772,9 +1787,12 @@ sub purchase_order_view {
 		} else {
 			$param{'delivered_on'} = undef;
 		} # end if
+		if ( ( $param{'vendor_country'} ne $PO->vendor_country() ) or ( $param{'vendor_state'} ne $PO->vendor_state() ) ) {
+			$PO->Taxes(1);
+		} # end if need to change taxes
 		foreach my $Tax ( $PO->Taxes() ) {
-			# Order is important here.
-			$Tax->charge($param{'tax_charge-'.$Tax->id()});
+			# Order is important here. Also the 1* turns an undef value into a specific boolean 0, because we used a checkbox
+			$Tax->charge(1*$param{'tax_charge-'.$Tax->id()}) if $Tax->charge() != 1*$param{'tax_charge-'.$Tax->id()};
 			$Tax->amount(undef);
 			$Tax->save();
 		} # end foreach
@@ -1816,21 +1834,81 @@ sub purchase_order_view {
 
 sub purchase_order_edit {
 
+	my $Me = new openprint::User( $session{'user_id'} );
 	my $PO = new openprint::PurchaseOrder( $param{'po_id'} );
 	
-	if ( $param{'btnFunction'} eq 'Delete' ) {
-		return;
-	} elsif ( $param{'btnFunction'} eq 'Save' ) {
+	if ( $param{'btnFunction'} eq 'Save' ) {
+		if ( ! $param{'po_id'} ) {
+			$variable{'error'} .= $PO->save( { 'created_by'	=>	$session{'user_id'}, 'company_id'=>$Me->company_id() } );
+		} # end if
+		if ( ! $param{'supplier_id'} ) {
+			my @Companies = openprint::Company->find( 'name'=>$param{'vendor_name'} );
+			if ( ! @Companies ) {
+				my $C = new openprint::Company();
+				$C->save({
+						'supplier'		=> 'Y',
+						'name'			=> $param{'vendor_name'},
+						'business_name'	=> $param{'vendor_name'},
+						'address1'		=> $param{'vendor_address1'},
+						'address2'		=> $param{'vendor_address2'},
+						'city'			=> $param{'vendor_city'},
+						'state'			=> $param{'vendor_state'},
+						'country'		=> $param{'vendor_country'},
+						'postalcode'	=> $param{'vendor_postalcode'},
+						'phone'			=> $param{'vendor_phone'},
+						'fax'			=> $param{'vendor_fax'},
+						} );
+				$param{'supplier_id'} = $C->id();
+			} else {
+				my $Company;
+				foreach my $C ( @Companies ) {
+					if ( $C->supplier() eq 'Y' ) {
+						$Company = $C;
+						last;
+					} # end if
+				} # end foreach
+				if ( ! $Company ) {
+					$Company = $Companies[0];
+					$Company->save( {'supplier'=>'Y'} );
+				} # end if
+				$param{'supplier_id'} = $Company->id();
+			} # end if
+		} # end if
+		if ( ! $param{'contact_id'} ) {
+			my @Users = openprint::User->find( 'company_id'=>$param{'supplier_id'}, 'email'=> lc $param{'vendor_email'} );
+			if ( ! @Users ) {
+				my $User = new openprint::User();
+				my ( $first, $last ) = $param{'vendor_contact'} =~ /(\S+)\s*(\S*)/;
+				$User->save( {
+						'company_id'=>	$param{'supplier_id'},
+						'email'		=>	$param{'vendor_email'},
+						'firstname'	=>	$first,
+						'lastname'	=>	$last,
+						'phone'		=>	$param{'vendor_phone'},
+						'fax'		=>	$param{'vendor_fax'},
+						'sms'		=>	$param{'vendor_sms'},
+						'change_password'	=>	'N',
+						'administrator'	=>	'N',
+						'ftp_active'	=>	0,
+						'web_active'	=>	0,
+					} );
+			} # end if
+		} # end if
+		if ( $param{'delivered_on_switch'} eq 'DATE' ) {
+			$param{'delivered_on'} = sprintf('%.4d-%.2d-%.2d', @param{'delivered_on_year','delivered_on_month','delivered_on_day'}) if ! $param{'delivered_on'};
+		} else {
+			$param{'delivered_on'} = undef;
+		} # end if
+		$variable{'error'} .= $PO->save( \%param );
 	} # end if btnFunction
 
 	if ( ! $PO->id() ) {
-		my $U = new openprint::User( $session{'user_id'} );
-		my $C = $U->Company();
+		my $C = $Me->Company();
 		$PO->set( {
 			'currency_id'		=>	openprint::Currency::get_current()->id(),
 			'company_id'		=>	$C->id(),
-			'created_by'		=>	$U->id(),
-			'shipto_contact'	=>	$U->name(),
+			'created_by'		=>	$Me->id(),
+			'shipto_contact'	=>	$Me->name(),
 			'shipto_name'		=>	$C->name(),
 			'shipto_address1'	=>	$C->address1(),
 			'shipto_address2'	=>	$C->address2(),
@@ -1839,12 +1917,11 @@ sub purchase_order_edit {
 			'shipto_country'	=>	$C->country(),
 			'shipto_postalcode'	=>	$C->postalcode(),
 			'shipto_phone'		=>	$C->phone(),
-			'shipto_mobile'		=>	$U->mobile(),
+			'shipto_mobile'		=>	$Me->mobile(),
 			'shipto_fax'		=>	$C->fax(),
-			'shipto_email'		=>	$U->email(),
-			'shipto_sms'		=>	$U->sms(),
+			'shipto_email'		=>	$Me->email(),
+			'shipto_sms'		=>	$Me->sms(),
 		} );
-		$variable{'error'} .= $PO->save();
 	} # end if
 	$variable{'PurchaseOrder'} = $PO;
 } # end sub purchase_order_edit
@@ -2035,5 +2112,7 @@ sub _skids_results {
 				) );
 }
 
+sub _update_taxes {
+} # end sub _update_taxes
 1;
 __END__
